ya esta perro

+++ DTOs/CreateSaleDto.cs
@@ -12,7 +12,8 @@ namespace ForrajeriaJovitaAPI.DTOs
     {
         public int ProductId { get; set; }
         public int ProductUnitId { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
+
         public decimal UnitPrice { get; set; }
         public decimal Discount { get; set; }
     }
